@@ -1,3 +1,3 @@
 /// GENERATED BY pub_release do not modify.
-/// future_builder_ex version
-String packageVersion = '3.0.0';
+/// Instance of 'Name' version
+String packageVersion = '3.1.0';
